ConservationLawsParticles.jl

Documentation for ConservationLawsParticles.jl.

Particle method for 1D conservation laws.

This package implements the deterministic particle schemes described in the article Entropy solutions of non-local scalar conservation laws with congestion via deterministic particle method, E. Radici, F. Stra (2021), https://arxiv.org/abs/2107.10760.

You can cite the article as

@online{RadiciStra2021,
    title={Entropy solutions of non-local scalar conservation laws with congestion via deterministic particle method}, 
    author={Emanuela Radici and Federico Stra},
    year={2021},
    eprint={2107.10760},
    archivePrefix={arXiv},
    primaryClass={math.AP},
    url={https://arxiv.org/abs/2107.10760}
}

Library

ConservationLawsParticles.AbstractModel

ConservationLawsParticles.IntegratedModel

ConservationLawsParticles.SampledModel

ConservationLawsParticles.eachspecies

ConservationLawsParticles.empty_like

ConservationLawsParticles.external_velocity

ConservationLawsParticles.integrated_interaction

ConservationLawsParticles.interaction

ConservationLawsParticles.make_velocity

ConservationLawsParticles.mobility

ConservationLawsParticles.num_species

ConservationLawsParticles.pwc_densities

ConservationLawsParticles.pwc_density

ConservationLawsParticles.pwc_density

ConservationLawsParticles.sampled_interaction

ConservationLawsParticles.species

ConservationLawsParticles.@time_independent

Public

ConservationLawsParticles.AbstractModel — Type

abstract type AbstractModel

Abstract type representing a particle model.

ConservationLawsParticles.IntegratedModel — Type

IntegratedModel((V₁, ...), ((W₁₁, ...), ...), (mob₁, ...)

Represents a particles system with:

external velocities Vᵢ,

integrated interactions Wᵢⱼ (this is the effect of the species j on the species i),

mobilities mobᵢ.

See also SampledModel.

Examples

julia> using ConservationLawsParticles.Examples, RecursiveArrayTools

julia> model = IntegratedModel(
           (V, V),
           ((W_attr, W_rep),
            (W_rep, W_attr)),
           (mobρ, mobσ));

julia> x = ArrayPartition(gaussian_particles(2, 4), collect(range(-3, 4, length=5)))
([-2.000000000000003, -0.30305473018369145, 0.30305473018369145, 2.000000000000003], [-3.0, -1.25, 0.5, 2.25, 4.0])

julia> velocities(x, model, 0.)
([6.621647425385332, 0.30545649966452776, -0.42671180044513507, -6.914302510772401], [23.261098611816987, 0.9023583690557855, 0.7055593913708742, -8.774526834146023, -55.23308442021724])

ConservationLawsParticles.SampledModel — Type

SampledModel((V₁, ...), ((W′₁₁, ...), ...), (mob₁, ...)

Represents a particles system with:

external velocities Vᵢ,

sampled interactions W′ᵢⱼ (this is the effect of the species j on the species i),

mobilities mobᵢ.

See also IntegratedModel.

Examples

julia> using ConservationLawsParticles.Examples, RecursiveArrayTools

julia> model = SampledModel(
           (V, V),
           ((Wprime_attr, Wprime_rep),
            (Wprime_rep, Wprime_attr)),
           (mobρ, mobσ));

julia> x = ArrayPartition(gaussian_particles(2, 4), collect(range(-3, 4, length=5)))
([-2.000000000000003, -0.30305473018369145, 0.30305473018369145, 2.000000000000003], [-3.0, -1.25, 0.5, 2.25, 4.0])

julia> velocities(x, model, 0.)
([6.291136247066298, 0.2466663161150116, -0.7218917091339228, -7.22630670503873], [22.405129478914613, 1.1249366684885518, 1.5188519354999799, -7.87111869358889, -54.536397957423915])

ConservationLawsParticles.eachspecies — Method

eachspecies(mod::AbstractModel) -> Any

Returns an iterator over the indices of the species of the model.

ConservationLawsParticles.empty_like — Function

empty_like(x::AbstractArray)

Creates an empty array with the same size and eltype of x.

Examples

julia> empty_like([1 2 3; 4 5 6])
2×3 Array{Int64,2}:
 0  140456850103040                1
 1                3  140456850103041

ConservationLawsParticles.external_velocity — Method

external_velocity(mod::AbstractModel, i::Integer) -> Any

Returns the external velocity field associated to the species i.

ConservationLawsParticles.integrated_interaction — Function

integrated_interaction([t,] x, W, ys[, dens_diff])

Computes $-(W' * \rho)(t, x) = -(W * \rho')(t,x)$, where $\rho$ is the piecewise-constant density associated to the particles ys.

It t is omitted, then W(x) is assumed independent of time.

Note

To ensure the correctness of the computation, dens_diff must coincide with diff(pwc_density(ys)). It can be pre-computed and passed explicitly to allow reuse (as an optimization).

See also sampled_interaction.

ConservationLawsParticles.interaction — Method

interaction(mod::AbstractModel, i::Integer, j::Integer) -> Any

Returns the interaction exerted on the species i by the species j.

ConservationLawsParticles.mobility — Method

mobility(mod::AbstractModel, i::Integer) -> Any

Returns the mobility associated to the species i.

ConservationLawsParticles.num_species — Method

num_species(mod::AbstractModel) -> Any

Returns the number of species of the model.

ConservationLawsParticles.pwc_densities — Method

pwc_densities(xs::AbstractVector...)

ConservationLawsParticles.pwc_density — Method

pwc_density(x::AbstractVector, y::AbstractVector)

The returned x_dens is indexed as x_dens[x_or_y, left_or_right, i].

ConservationLawsParticles.pwc_density — Method

pwc_density(x::AbstractVector)

Returns the piecewise constant probability density from the quantile particle positions.

Let the quantile particles be at positions x₀, x₁, …, xₙ (remember that in Julia they are actually x[1], x[2], ..., x[n], x[n+1]). The returned array R is such that R[1] = R[n+2] = 0 and R[i] = 1 / (N * (x[i] - x[i-1])) for the intermediate indices (this formula holds also for the first and last entry if we assume x[0] = -∞ and x[n+2] = ∞).

Examples

julia> pwc_density([0, 1, 3])
4-element Vector{Float64}:
 0.0
 0.5
 0.25
 0.0

ConservationLawsParticles.sampled_interaction — Function

sampled_interaction([t,] x, W′, ys)

Computes $-(W' * \dot\rho)(t, x)$, which is the sampled approximation of $-(W' * \rho)(t, x)$, where $\rho$ is the piecewise-constant density associated to the particles ys.

It t is omitted, then W′(x) is assumed independent of time.

See also integrated_interaction.

ConservationLawsParticles.species — Method

species(state, i::Integer) -> Any

Returns the particles associated to the species i.

ConservationLawsParticles.@time_independent — Macro

Automatically define a method which takes time as first argument and discards it.

Examples

The definition

@time_independent V(x) = -x^3

is equivalent to

V(x) = -x^3
V(t, x) = V(x)

This also works with more than one argument, for instance

@time_independent V(x₁, x₂) = x₁ * x₂

is equivalent to

V(x₁, x₂) = x₁ * x₂
V(t, x₁, x₂) = V(x₁, x₂)

Private

ConservationLawsParticles.make_velocity — Method

make_velocity(V::Function, Wprime::Function, mobility::Function)

Creates a function velocity(dx, x, p, t) that computes the velocity of the particles under the influence of an external force V, a mutual interaction Wprime and the congestion given by mobility.